python/run/hello-auditlog/requirements.txt
flask==2.1.0
cloudevents==1.9.0
google-events==0.4.0

